+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/model/domain/SysUser.java
@@ -40,7 +40,7 @@ public class SysUser extends BaseModel<SysUser> {
     /**
      * 所属租户
      */
-    private Long tenantId;
+    private Integer tenantId;
 
     /**
      * 部门ID
